\docType{data}
\name{detection}
\alias{detection}
\title{Data from a detection experiment.}
\description{
  This data file can be used for testing fitting of a
  within-subjects design: performance as a function of
  contrast for four subjects.
}
\details{
  This contains some trials for detecting a target
  presented in one of two temporal intervals (i.e. 2AFC).
  Four subjects completed a variable number of trials.

  The \code{contrast} column contains the Weber contrast of
  the target. The \code{interval} column is whether the
  target appeared in the first or second interval. The
  \code{correct} column is a binary value for whether the
  subject answered the trial correctly. The \code{subject}
  column is a string containing the subject code.
}
\author{
  Thomas Wallis \email{thomas.wallis@uni-tuebingen.de}
}
\keyword{datasets}

