\docType{package}
\name{psybayes}
\alias{package-psybayes}
\alias{psybayes}
\alias{psybayes-package}
\title{Fitting psychophysical data with MCMC.}
\description{
  psybayes is a package to facilitate the fitting of data
  from psychophysical experiments in a Bayesian framework.
  Models are fit using MCMC sampling conducted through Stan
  (and the rstan package). These will need to be properly
  installed on your system first (see
  \url{http://www.mc-stan.org} for setup). Many of these
  functions will end up being wrappers for rstan calls, or
  plotting functions that are commonly used.
}
\details{
  psybayes

  Package: psybayes Type: package Version: 0.1 License:
  GPL-3
}
\author{
  Thomas Wallis \email{thomas.wallis@uni-tuebingen.de}

  includes some code written by John Kruschke for the book
  Doing Bayesian Data Analysis.
}

